+++ src/assumptions.jl
@@ -86,20 +86,4 @@ function parse_type_condition(f::Union{Type{Val{:<}}, Type{Val{:>}}}, condition)
     end
 end
 
-# isless
-
-isless(x, y) = Base.isless(x, y)
-
-# commutes
-
-commutes(x, y) = false
-commutes(::Number, ::Number) = true
-commutes(::Number, ::Any) = true
-commutes(::Any, ::Number) = false
-
-# sort
-
-add_order(x, y) = isless(string(firstsymbol(x, x)), string(firstsymbol(y, y)))
-mul_order(x, y) = commutes(x, y)
-
 end
